app/requirements.txt
flask==2.3.2
prometheus_client==0.17.0
python-json-logger==2.0.7
requests==2.31.0

opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-propagator-b3==1.21.0
opentelemetry-exporter-jaeger==1.21.0

